Added maze generation EditorWindow
In the editor menu, select Custom > Maze Generation. Enter a maze size, drag and drop a floor and wall GameObject, and click Generate to generate a maze. Click Destroy to destroy a generated maze. I plan to extract the editor window functionality from this class to a new dedicated MazeGeneratorWindow class.